% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{helloYou}
\alias{helloYou}
\title{helloYou}
\usage{
helloYou(user)
}
\arguments{
\item{user}{}
}
\value{
character string that says hello to user from world
}
\description{
user puts there name in and this fxn says hello to them
}
\examples{
helloYou("Bruce Wayne")

}
