% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{maf}
\alias{maf}
\title{maf}
\usage{
maf(a, b)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}
}
\value{
c is a numeric of any soft of type depending on what user puts in. In fact maybe enven a string woudlwrk!
}
\description{
this adds to gether a and b using the + operator. What does it do wiht vec, str, or dfs?
}
\examples{
maf(1,2)
maf(1999,1)
}
